MemoryPool: Added a feature and more tests
Added MemoryPool::segmentList() to allow inspection of the current memory
layout in the MemoryPool.  Not only does this allow for more fine-tuned
testing but it will enable the ability for applications to visualize the
memory usage.  MemoryPool::dump() already does this to a very limited
extent.

Also added a to_string() to convert the segment data into a JSON string.

The "Reuse" tests use the new segmentList() to verify that memory allocation
is happening in the expected areas of the memory pool.  These tests will
need to be reworked if/when the allocation strategy becomes configurable.

FossilOrigin-Name: d086a36d1489e04f479429df0c8d60c70a6635edd522e2af372707790d991cea